# api/openapi.yaml
openapi: 3.1.0
info:
  title: audit-in-a-box API
  version: 0.2.0
servers:
  - url: http://localhost:25342
paths:
  /healthz:
    get:
      responses:
        "200":
          description: Healthy
  /readyz:
    get:
      responses:
        "200":
          description: Ready with tool status
  /metrics:
    get:
      responses:
        "200":
          description: Prometheus metrics
  /api/v1/tools:
    get:
      responses:
        "200":
          description: Tool availability
  /api/v1/scrape:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [html]
              properties:
                html:
                  type: string
      responses:
        "200":
          description: Manifest candidates
        "400":
          description: Recoverable scrape error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/audits:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuditRequest"
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                file_name:
                  type: string
                ecosystem:
                  type: string
                pasted_html:
                  type: string
      responses:
        "200":
          description: Audit report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditReport"
        "422":
          description: Recoverable audit error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    AuditRequest:
      type: object
      properties:
        file_name:
          type: string
        content:
          type: string
        pasted_html:
          type: string
        ecosystem:
          type: string
          enum: [npm, go, python, unknown]
    AuditReport:
      type: object
      required: [id, generated_at, input, provenance, dependencies, vulnerabilities, license_risks, anomalies, maintainer_health, risk, summary]
      properties:
        id:
          type: string
        generated_at:
          type: string
          format: date-time
        summary:
          type: string
        input:
          $ref: "#/components/schemas/AuditInput"
        provenance:
          type: object
          properties:
            schema_version:
              type: string
            source_hash:
              type: string
            parameters:
              type: object
              additionalProperties: true
        risk:
          type: object
          properties:
            score:
              type: integer
            grade:
              type: string
            confidence:
              type: number
            counts:
              type: object
              additionalProperties:
                type: integer
            factors:
              type: array
              items:
                type: string
        dependencies:
          type: array
          items:
            $ref: "#/components/schemas/Dependency"
        vulnerabilities:
          type: array
          items:
            $ref: "#/components/schemas/Vulnerability"
        license_risks:
          type: array
          items:
            $ref: "#/components/schemas/LicenseRisk"
        anomalies:
          type: array
          items:
            $ref: "#/components/schemas/ReportAnomaly"
        maintainer_health:
          type: array
          items:
            type: object
    AuditInput:
      type: object
      properties:
        file_name:
          type: string
        ecosystem:
          type: string
        kind:
          type: string
        parser:
          type: string
        confidence:
          type: number
        source_hash:
          type: string
    Dependency:
      type: object
      properties:
        name:
          type: string
        version:
          type: string
        ecosystem:
          type: string
        scope:
          type: string
        licenses:
          type: array
          items:
            type: string
        source:
          type: string
        confidence:
          type: number
        reasons:
          type: array
          items:
            type: string
    Vulnerability:
      type: object
      properties:
        id:
          type: string
        package_name:
          type: string
        severity:
          type: string
        confidence:
          type: number
    LicenseRisk:
      type: object
      properties:
        package_name:
          type: string
        license:
          type: string
        severity:
          type: string
        reason:
          type: string
        confidence:
          type: number
    ReportAnomaly:
      type: object
      properties:
        code:
          type: string
        severity:
          type: string
        message:
          type: string
        why:
          type: string
        next_step:
          type: string
        confidence:
          type: number
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            why:
              type: string
            next_step:
              type: string
            recoverable:
              type: boolean